Add usesize for size logic moving to clean up size component
